+++ src/shapes/Quadrilateral.java
@@ -1,21 +1,23 @@
 package shapes;
 
 public abstract class Quadrilateral extends Shape implements Measurable{
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
-    public int getWidth() {
+    // getters
+    public double getWidth() {
         return width;
     }
-    public int getLength() {
+    public double getLength() {
         return length;
     }
 
-    public abstract void setLength(int length);
-    public abstract void setWidth(int width);
+    // setters
+    public abstract void setLength(double length);
+    public abstract void setWidth(double width);
 
-
-    public Quadrilateral(int length, int width) {
+    // constructors
+    public Quadrilateral(double length, double width) {
         this.length = length;
         this.width = width;
     }
